requirements.txt
Django==5.2.4
asgiref==3.9.1
sqlparse==0.5.3
requests==2.32.4
python-dotenv==1.1.1
psycopg2-binary==2.9.10
gunicorn==23.0.0
whitenoise==6.6.0
dj-database-url==2.1.0
mecab-python3==1.0.10
konlpy==0.6.0
django-sslserver==0.22
Pillow==10.4.0
pillow-heif==0.18.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
grpcio==1.65.5
grpcio-tools==1.65.5
protobuf==5.27.2
sounddevice==0.5.0
fastapi==0.116.1
uvicorn[standard]==0.35.0
websockets==15.0.1
